fix(date): some posts had different dates in cms and generated site
It happened because the datetime widget in the CMS was saving to Github the date in
UTC which was then displayed without a conversion back to local time.
Thus, when a time was close to midnight (e.g. 11:00 pm) then the
date saved was the day after since UTC adds hours (compared to the New York
timezone). It was then displayed as UTC, thus the website was
showing a day later than correct.